Refactor Eloverblik integration to utilize a centralized normalization function for input data, enhancing validation by checking for empty fields. Update version to 0.7.2 and add error messages for empty fields in multiple languages.
