+++ custom_components/eloverblik/manifest.json
@@ -1,7 +1,7 @@
 {
   "domain": "eloverblik",
   "name": "Eloverblik",
-  "version": "0.7.1",
+  "version": "0.7.2",
   "config_flow": true,
   "documentation": "https://github.com/JonasPed/homeassistant-eloverblik",
   "issue_tracker": "https://github.com/JonasPed/homeassistant-eloverblik/issues",
